+++ crates/teloxide-core/src/payloads.rs
@@ -87,6 +87,7 @@ mod get_my_default_administrator_rights;
 mod get_my_description;
 mod get_my_name;
 mod get_my_short_description;
+mod get_my_star_balance;
 mod get_star_transactions;
 mod get_sticker_set;
 mod get_updates;
@@ -267,6 +268,7 @@ pub use get_my_default_administrator_rights::{
 pub use get_my_description::{GetMyDescription, GetMyDescriptionSetters};
 pub use get_my_name::{GetMyName, GetMyNameSetters};
 pub use get_my_short_description::{GetMyShortDescription, GetMyShortDescriptionSetters};
+pub use get_my_star_balance::{GetMyStarBalance, GetMyStarBalanceSetters};
 pub use get_star_transactions::{GetStarTransactions, GetStarTransactionsSetters};
 pub use get_sticker_set::{GetStickerSet, GetStickerSetSetters};
 pub use get_updates::{GetUpdates, GetUpdatesSetters};